PKINIT testing: don't always expect responder questions
When testing authenticated (non-anonymous) PKINIT, expect to get a
challenge from PKINIT only when there's a smart card that requires login
or there's a private key that needs to be decrypted.  Add PKCS11 tests
which use an unencrypted key, and check that we also don't produce a
challenge question when using it.