Perform complete referrals in t_referral.py
t_referral.py was written to exercise KDC host referral logic, and did
not actually create the target realm, instead just looking at the
error message from gcred to determine whether the KDC returned a
referral or not.  It's only a small amount of additional work to
actually set up the target realm and check that the client code
successfully retrieves the referral, so do that instead.

Since the referral and non-referral outcomes aren't all that similar
any more, split test() into testref() and testfail().  Get rid of the
message argument, since it wouldn't be output in most cases where we
get an unexpected result.