Import MSSQLSvc/host:port as host-based service name
Sadly Microsoft's principal naming scheme for SQL server instances
is not GSSAPI friendly.  Instead of encoding the port number into
the service name, they append it to the hostname:

    MSSQLSvc/db.example.com:1433

At least some JDBC libraries specify server name of
"MSSQLSvc/db.example.com:1433" with a null name type.

This only works when the database server is in the default realm
and is not an alias.  If either condition is false, the client is
unable to obtain the required tickets.

We apply the mssql workaround for two component null and kerberos
names with a service part of MSSQLSvc.  This can be disabled via a
new libdefaults boolean (mssql_workaround = true|false).