Support referrals from Windows Server 2003
Although the referrals draft requires servers to indicate a referral
in a WRONG_REALM message, Microsoft Windows Server 2003 returns this
information in a message with error code PRINCIPAL_UNKNOWN. Failure to
follow the referral in these messages prevents referral chasing in
Windows Server 2003 forests. This patch detects referral messages of
this type by checking for a non-empty client.realm field in the
response, and activate the referral logic in these cases.