Remove MBCS from leash's DEFINES
This is just enabling the use of multi-byte character set in the
MFC library, but we do not appear to make use of this feature.
Visual Studio 2013 gives ominous warnings that support for it may
be removed in future versions, so quiet the build and do not
enable the deprecated feature we are not using.

(cherry picked from commit 5d50bc2a62739e39ea0b90b0ced3b7489ba554b3)

ticket: 8190
version_fixed: 1.13.3