Fix memory leak in krb5_verify_init_creds()
When copying most of the credentials from one cache to another in
copy_creds_except() (called from get_vfy_cred(), from
krb5_verify_init_creds()), we need to explicitly free all of the
credentials that we read.  We previously didn't free the ones which we
didn't store in the destination cache.